# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Monitor API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Account {
  id: string;
  label: string;
  createdAt: string;
}

export interface CreateAccountBody {
  label: string;
}

export type WhatsappStatusState =
  (typeof WhatsappStatusState)[keyof typeof WhatsappStatusState];

export const WhatsappStatusState = {
  INITIALIZING: "INITIALIZING",
  QR_READY: "QR_READY",
  AUTHENTICATED: "AUTHENTICATED",
  READY: "READY",
  DISCONNECTED: "DISCONNECTED",
} as const;

export interface WhatsappStatus {
  state: WhatsappStatusState;
  phoneNumber?: string | null;
  displayName?: string | null;
  profilePicUrl?: string | null;
}

export interface QrCode {
  /** QR code data string */
  qr: string;
  /** Base64 data URL of QR code image */
  qrDataUrl?: string | null;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface Chat {
  id: string;
  name: string;
  isGroup: boolean;
  lastMessage?: string | null;
  lastMessageTimestamp?: number | null;
  unreadCount: number;
  profilePicUrl?: string | null;
  isArchived: boolean;
  isMuted: boolean;
}

export interface Message {
  id: string;
  body: string;
  timestamp: number;
  fromMe: boolean;
  author?: string | null;
  type: string;
  hasMedia: boolean;
  isForwarded: boolean;
  isStarred: boolean;
}

export interface WhatsappStats {
  totalChats: number;
  totalGroups: number;
  totalPrivateChats: number;
  unreadChats: number;
  totalUnreadMessages: number;
}

export type GetAccountChatMessagesParams = {
  limit?: number;
};