# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: WhatsApp Monitor API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: whatsapp
    description: WhatsApp operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /whatsapp/accounts:
    get:
      operationId: listAccounts
      tags: [whatsapp]
      summary: List all linked WhatsApp accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [whatsapp]
      summary: Add a new WhatsApp account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"

  /whatsapp/accounts/{accountId}:
    delete:
      operationId: deleteAccount
      tags: [whatsapp]
      summary: Remove a WhatsApp account and clear its session
      parameters:
        - $ref: "#/components/parameters/accountId"
      responses:
        "200":
          description: Account removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /whatsapp/accounts/{accountId}/status:
    get:
      operationId: getAccountStatus
      tags: [whatsapp]
      summary: Get WhatsApp connection status for an account
      parameters:
        - $ref: "#/components/parameters/accountId"
      responses:
        "200":
          description: Connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappStatus"
        "404":
          description: Account not found

  /whatsapp/accounts/{accountId}/qr:
    get:
      operationId: getAccountQr
      tags: [whatsapp]
      summary: Get QR code for linking an account
      parameters:
        - $ref: "#/components/parameters/accountId"
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
        "404":
          description: No QR code available yet

  /whatsapp/accounts/{accountId}/logout:
    post:
      operationId: logoutAccount
      tags: [whatsapp]
      summary: Logout / unlink a WhatsApp account (keeps the account slot)
      parameters:
        - $ref: "#/components/parameters/accountId"
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /whatsapp/accounts/{accountId}/chats:
    get:
      operationId: getAccountChats
      tags: [whatsapp]
      summary: Get all chats for an account
      parameters:
        - $ref: "#/components/parameters/accountId"
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"

  /whatsapp/accounts/{accountId}/chats/{chatId}/messages:
    get:
      operationId: getAccountChatMessages
      tags: [whatsapp]
      summary: Get messages for a chat within an account
      parameters:
        - $ref: "#/components/parameters/accountId"
        - name: chatId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  /whatsapp/accounts/{accountId}/stats:
    get:
      operationId: getAccountStats
      tags: [whatsapp]
      summary: Get summary stats for an account
      parameters:
        - $ref: "#/components/parameters/accountId"
      responses:
        "200":
          description: Stats overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappStats"

components:
  parameters:
    accountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Account:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        createdAt:
          type: string
      required:
        - id
        - label
        - createdAt

    CreateAccountBody:
      type: object
      properties:
        label:
          type: string
      required:
        - label

    WhatsappStatus:
      type: object
      properties:
        state:
          type: string
          enum: [INITIALIZING, QR_READY, AUTHENTICATED, READY, DISCONNECTED]
        phoneNumber:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        profilePicUrl:
          type: string
          nullable: true
      required:
        - state

    QrCode:
      type: object
      properties:
        qr:
          type: string
          description: QR code data string
        qrDataUrl:
          type: string
          description: Base64 data URL of QR code image
          nullable: true
      required:
        - qr

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    Chat:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        isGroup:
          type: boolean
        lastMessage:
          type: string
          nullable: true
        lastMessageTimestamp:
          type: integer
          nullable: true
        unreadCount:
          type: integer
        profilePicUrl:
          type: string
          nullable: true
        isArchived:
          type: boolean
        isMuted:
          type: boolean
      required:
        - id
        - name
        - isGroup
        - unreadCount
        - isArchived
        - isMuted

    Message:
      type: object
      properties:
        id:
          type: string
        body:
          type: string
        timestamp:
          type: integer
        fromMe:
          type: boolean
        author:
          type: string
          nullable: true
        type:
          type: string
        hasMedia:
          type: boolean
        isForwarded:
          type: boolean
        isStarred:
          type: boolean
      required:
        - id
        - body
        - timestamp
        - fromMe
        - type
        - hasMedia
        - isForwarded
        - isStarred

    WhatsappStats:
      type: object
      properties:
        totalChats:
          type: integer
        totalGroups:
          type: integer
        totalPrivateChats:
          type: integer
        unreadChats:
          type: integer
        totalUnreadMessages:
          type: integer
      required:
        - totalChats
        - totalGroups
        - totalPrivateChats
        - unreadChats
        - totalUnreadMessages