# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Monitor API
 * OpenAPI spec version: 0.2.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all linked WhatsApp accounts
 */
export const ListAccountsResponseItem = zod.object({
  id: zod.string(),
  label: zod.string(),
  createdAt: zod.string(),
});
export const ListAccountsResponse = zod.array(ListAccountsResponseItem);

/**
 * @summary Add a new WhatsApp account
 */
export const CreateAccountBody = zod.object({
  label: zod.string(),
});

/**
 * @summary Remove a WhatsApp account and clear its session
 */
export const DeleteAccountParams = zod.object({
  accountId: zod.coerce.string(),
});

export const DeleteAccountResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get WhatsApp connection status for an account
 */
export const GetAccountStatusParams = zod.object({
  accountId: zod.coerce.string(),
});

export const GetAccountStatusResponse = zod.object({
  state: zod.enum([
    "INITIALIZING",
    "QR_READY",
    "AUTHENTICATED",
    "READY",
    "DISCONNECTED",
  ]),
  phoneNumber: zod.string().nullish(),
  displayName: zod.string().nullish(),
  profilePicUrl: zod.string().nullish(),
});

/**
 * @summary Get QR code for linking an account
 */
export const GetAccountQrParams = zod.object({
  accountId: zod.coerce.string(),
});

export const GetAccountQrResponse = zod.object({
  qr: zod.string().describe("QR code data string"),
  qrDataUrl: zod
    .string()
    .nullish()
    .describe("Base64 data URL of QR code image"),
});

/**
 * @summary Logout / unlink a WhatsApp account (keeps the account slot)
 */
export const LogoutAccountParams = zod.object({
  accountId: zod.coerce.string(),
});

export const LogoutAccountResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get all chats for an account
 */
export const GetAccountChatsParams = zod.object({
  accountId: zod.coerce.string(),
});

export const GetAccountChatsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  isGroup: zod.boolean(),
  lastMessage: zod.string().nullish(),
  lastMessageTimestamp: zod.number().nullish(),
  unreadCount: zod.number(),
  profilePicUrl: zod.string().nullish(),
  isArchived: zod.boolean(),
  isMuted: zod.boolean(),
});
export const GetAccountChatsResponse = zod.array(GetAccountChatsResponseItem);

/**
 * @summary Get messages for a chat within an account
 */
export const GetAccountChatMessagesParams = zod.object({
  accountId: zod.coerce.string(),
  chatId: zod.coerce.string(),
});

export const getAccountChatMessagesQueryLimitDefault = 50;

export const GetAccountChatMessagesQueryParams = zod.object({
  limit: zod.coerce.number().default(getAccountChatMessagesQueryLimitDefault),
});

export const GetAccountChatMessagesResponseItem = zod.object({
  id: zod.string(),
  body: zod.string(),
  timestamp: zod.number(),
  fromMe: zod.boolean(),
  author: zod.string().nullish(),
  type: zod.string(),
  hasMedia: zod.boolean(),
  isForwarded: zod.boolean(),
  isStarred: zod.boolean(),
});
export const GetAccountChatMessagesResponse = zod.array(
  GetAccountChatMessagesResponseItem,
);

/**
 * @summary Get summary stats for an account
 */
export const GetAccountStatsParams = zod.object({
  accountId: zod.coerce.string(),
});

export const GetAccountStatsResponse = zod.object({
  totalChats: zod.number(),
  totalGroups: zod.number(),
  totalPrivateChats: zod.number(),
  unreadChats: zod.number(),
  totalUnreadMessages: zod.number(),
});