# lib/api-zod/src/generated/types/chat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Monitor API
 * OpenAPI spec version: 0.2.0
 */

export interface Chat {
  id: string;
  name: string;
  isGroup: boolean;
  lastMessage?: string | null;
  lastMessageTimestamp?: number | null;
  unreadCount: number;
  profilePicUrl?: string | null;
  isArchived: boolean;
  isMuted: boolean;
}