# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Monitor API
 * OpenAPI spec version: 0.2.0
 */

export interface Message {
  id: string;
  body: string;
  timestamp: number;
  fromMe: boolean;
  author?: string | null;
  type: string;
  hasMedia: boolean;
  isForwarded: boolean;
  isStarred: boolean;
}