# lib/api-zod/src/generated/types/qrCode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Monitor API
 * OpenAPI spec version: 0.2.0
 */

export interface QrCode {
  /** QR code data string */
  qr: string;
  /** Base64 data URL of QR code image */
  qrDataUrl?: string | null;
}