# lib/api-zod/src/generated/types/whatsappStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Monitor API
 * OpenAPI spec version: 0.2.0
 */

export interface WhatsappStats {
  totalChats: number;
  totalGroups: number;
  totalPrivateChats: number;
  unreadChats: number;
  totalUnreadMessages: number;
}