# lib/api-zod/src/generated/types/whatsappStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Monitor API
 * OpenAPI spec version: 0.2.0
 */
import type { WhatsappStatusState } from "./whatsappStatusState";

export interface WhatsappStatus {
  state: WhatsappStatusState;
  phoneNumber?: string | null;
  displayName?: string | null;
  profilePicUrl?: string | null;
}